// src/types/movie.interface.ts
export interface Movie {
  adult: boolean;
  backdrop_path: string | null;
  genre_ids: number[];
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string | null;
  release_date: string; // ISO date string (YYYY-MM-DD)
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
}